+++ src/interface_type_001_actuators.h
@@ -16,7 +16,7 @@
 struct actuator_s
 {
    char name[20];
-   int actuator_id;
+   int16_t actuator_id;
    
    char arduino_pin;
    char arduino_pin_type;
@@ -28,7 +28,7 @@ struct actuator_s
 };
 
 void _interface_type_001_free_actuators_queue_elem(void *d);
-struct actuator_s *valid_and_malloc_actuator(int id_sensor_actuator, char *name, char *parameters);
+struct actuator_s *valid_and_malloc_actuator(int16_t id_sensor_actuator, char *name, char *parameters);
 mea_error_t xpl_actuator(interface_type_001_t *i001, xPL_NameValueListPtr ListNomsValeursPtr, char *device, char *type);
 
 #endif
